object/verb/property flag/preposition fields are 16 bits
prepositions need to be signed because of PREP_ANY and PREP_NONE
but everything else can be unsigned.

+++ db_private.h
@@ -34,8 +34,8 @@ struct Verbdef {
     const char *name;
     Program *program;
     Objid owner;
-    short perms;
-    short prep;
+    uint16_t perms;
+    int16_t  prep;
     Verbdef *next;
 };
 
@@ -56,7 +56,7 @@ struct Proplist {
 typedef struct Pval {
     Var var;
     Objid owner;
-    short perms;
+    uint16_t perms;
 } Pval;
 
 typedef struct Object {
@@ -72,7 +72,7 @@ typedef struct Object {
 
 
     const char *name;
-    int flags;
+    uint16_t flags;
 
     Verbdef *verbdefs;
     Proplist propdefs;
